Move color picker to a separate module.
