+++ gradle/settings.substitutions.gradle
@@ -37,6 +37,7 @@ dependencySubstitution {
 	substitute module("net.twisterrob.lib:twister-lib-android-basics") with project(":basics")
 	substitute module("net.twisterrob.lib:twister-lib-android-capture_image") with project(":capture_image")
 	substitute module("net.twisterrob.lib:twister-lib-android-capture_image-test_helpers") with project(":capture_image-test_helpers")
+	substitute module("net.twisterrob.lib:twister-lib-android-color_picker") with project(":color_picker")
 	substitute module("net.twisterrob.lib:twister-lib-android-cpsuite") with project(":cpsuite")
 	substitute module("net.twisterrob.lib:twister-lib-android-defs") with project(":defs")
 	substitute module("net.twisterrob.lib:twister-lib-android-espresso") with project(":espresso")
